Split SQL test files - no code change
